requirements.txt
openai
python-dotenv
PyPDF2
playwright
tqdm
